style: fix eslint warnings
